Rename to `Test.Data.Store`
