+++ lib/delta-store/delta-store.cabal
@@ -60,7 +60,7 @@ library
     Data.DBVar
     Data.Delta.Update
     Data.Store
-    Test.Store
+    Test.Data.Store
 
 test-suite unit
   import:             language, opts-exe
